Translate LIETUVIU.java
Please contact www.reddit.com/u/peraz if I missed something or any questions regarding my translation :)

Also, if you will put it in the next update, the correct word for my language is "Lietuvių" :)
